Move rdfparser.js to rdfxmlparser.js
