+++ src/rdfxmlparser.js
